+++ packages/adapters/wagmi/src/client.ts
@@ -311,8 +311,12 @@ export class WagmiAdapter implements ChainAdapter {
           ReturnType<(typeof UniversalProvider)['init']>
         >
 
-        const siweParams = await this.options?.siweConfig?.getMessageParams?.()
+        const clientId = await provider?.client?.core?.crypto?.getClientId()
+        if (clientId) {
+          this.appKit?.setClientId(clientId)
+        }
 
+        const siweParams = await this.options?.siweConfig?.getMessageParams?.()
         const isSiweEnabled = this.options?.siweConfig?.options?.enabled
         const isProviderSupported = typeof provider?.authenticate === 'function'
         const isSiweParamsValid = siweParams && Object.keys(siweParams || {}).length > 0
@@ -935,6 +939,8 @@ export class WagmiAdapter implements ChainAdapter {
       // Auth connector is initialized separately
       const shouldSkip = ConstantsUtil.AUTH_CONNECTOR_ID === id
       if (!shouldSkip) {
+        const injectedConnector = id === ConstantsUtil.INJECTED_CONNECTOR_ID
+
         w3mConnectors.push({
           id,
           explorerId: PresetsUtil.ConnectorExplorerIds[id],
@@ -942,9 +948,7 @@ export class WagmiAdapter implements ChainAdapter {
           name: PresetsUtil.ConnectorNamesMap[id] ?? name,
           imageId: PresetsUtil.ConnectorImageIds[id],
           type: PresetsUtil.ConnectorTypesMap[type] ?? 'EXTERNAL',
-          info: {
-            rdns: id
-          },
+          info: injectedConnector ? undefined : { rdns: id },
           chain: this.chainNamespace
         })
       }