+++ packages/appkit/src/utils/HelpersUtil.ts
@@ -1,6 +1,7 @@
 import type { NamespaceConfig, Namespace } from '@walletconnect/universal-provider'
 import type { CaipNetwork, CaipNetworkId, ChainNamespace } from '@reown/appkit-common'
 import type { SessionTypes } from '@walletconnect/types'
+import { solana, solanaDevnet } from '../networks/index.js'
 
 export const WcHelpersUtil = {
   getMethodsByChainNamespace(chainNamespace: ChainNamespace): string[] {
@@ -59,6 +60,17 @@ export const WcHelpersUtil = {
       const namespace = acc[chainNamespace] as Namespace
 
       namespace.chains.push(caipNetworkId)
+
+      // Workaround for wallets that only support deprecated Solana network ID
+      switch (caipNetworkId) {
+        case solana.caipNetworkId:
+          namespace.chains.push(solana.deprecatedCaipNetworkId)
+          break
+        case solanaDevnet.deprecatedCaipNetworkId:
+          namespace.chains.push(solanaDevnet.deprecatedCaipNetworkId)
+          break
+        default:
+      }
 
       if (namespace?.rpcMap && rpcUrl) {
         namespace.rpcMap[id] = rpcUrl