+++ packages/scaffold-ui/src/views/w3m-onramp-tokens-select-view/styles.ts
@@ -6,7 +6,18 @@ export default css`
     overflow: auto;
   }
 
+  wui-flex {
+    transition: opacity var(--wui-ease-out-power-1) var(--wui-duration-md);
+    will-change: opacity;
+  }
+
   wui-grid::-webkit-scrollbar {
     display: none;
+  }
+
+  wui-flex.disabled {
+    opacity: 0.3;
+    pointer-events: none;
+    user-select: none;
   }
 `